initial refactor of names and files
