+++ .gitignore
@@ -5,3 +5,4 @@ go.sum
 example/example
 *.old
 cmd/geneosctl/geneosctl
+cmd/geneos/geneos
