+++ cmd/geneos/README.md
