+++ cmd/geneos/command_reload.go
@@ -1,0 +1,27 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"syscall"
+)
+
+func Reload(c Component) (err error) {
+	pid, err := findProc(c)
+	if err != nil {
+		return
+	}
+
+	if !canControl(c) {
+		// fail early
+		return fmt.Errorf("cannot control process")
+	}
+
+	// send a SIGUSR1
+	proc, _ := os.FindProcess(pid)
+	if err := proc.Signal(syscall.SIGUSR1); err != nil {
+		log.Println(Type(c), Name(c), "refresh failed")
+
+	}
+	return
+}
