+++ cmd/geneos/component.go
