+++ cmd/geneos/gateway.go
