+++ cmd/geneos/licd.go
