+++ cmd/geneos/utils.go
@@ -15,124 +15,9 @@ import (
 	"strconv"
 	"strings"
 	"syscall"
-	"time"
 )
 
 var itrsHome string = "/opt/itrs"
-
-func Stop(c Component) (err error) {
-	pid, err := findProc(c)
-	if err != nil {
-		return
-	}
-
-	// who is the process running as?
-	s, _ := os.Stat(fmt.Sprintf("/proc/%d", pid))
-	st := s.Sys().(*syscall.Stat_t)
-	log.Println("process running as", st.Uid, st.Gid)
-
-	proc, _ := os.FindProcess(pid)
-
-	if canControl(c) {
-		log.Println("stopping", Type(c), Name(c), "PID", pid)
-
-		if err = proc.Signal(syscall.SIGTERM); err != nil {
-			log.Println("sending SIGTERM failed:", err)
-			return
-		}
-
-		// send a signal 0 in a loop to check if the process has terminated
-		for i := 0; i < 10; i++ {
-			time.Sleep(250 * time.Millisecond)
-			if err = proc.Signal(syscall.Signal(0)); err != nil {
-				log.Println(Type(c), "terminated")
-				return nil
-			}
-		}
-		// sigkill
-		if err = proc.Signal(syscall.SIGKILL); err != nil {
-			log.Println("sending SIGKILL failed:", err)
-			return
-		}
-	} else {
-		err = fmt.Errorf("cannot stop %s %s process, no permission (%v != %v)", Type(c), Name(c), st.Uid, os.Getuid())
-	}
-	return
-}
-
-func Start(c Component) (err error) {
-	cmd, env := Command(c)
-	if cmd == nil {
-		return
-	}
-
-	if !canControl(c) {
-		// fail early
-		return fmt.Errorf("cannot control process")
-	}
-
-	if superuser {
-		// set underlying user for child proc
-		username := getString(c, Prefix(c)+"User")
-		err = setuid(cmd, username)
-		if err != nil {
-			return
-		}
-	}
-
-	cmd.Env = append(os.Environ(), env...)
-
-	errfile := filepath.Join(getString(c, Prefix(c)+"LogD"), Type(c).String()+".txt")
-
-	out, err := os.OpenFile(errfile, os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		log.Fatalf("cannot open %q: %s\n", errfile, err)
-	}
-
-	if cmd.SysProcAttr != nil && superuser {
-		err = out.Chown(int(cmd.SysProcAttr.Credential.Uid), int(cmd.SysProcAttr.Credential.Gid))
-		if err != nil {
-			log.Println("chown:", err)
-		}
-	}
-	cmd.Stdout = out
-	cmd.Stderr = out
-	cmd.Dir = filepath.Join(Home(c))
-
-	err = cmd.Start()
-	if err != nil {
-		log.Println(err)
-		return
-	}
-	DebugLogger.Println("started process", cmd.Process.Pid)
-
-	if cmd.Process != nil {
-		// detach from control
-		cmd.Process.Release()
-	}
-
-	return
-}
-
-func Reload(c Component) (err error) {
-	pid, err := findProc(c)
-	if err != nil {
-		return
-	}
-
-	if !canControl(c) {
-		// fail early
-		return fmt.Errorf("cannot control process")
-	}
-
-	// send a SIGUSR1
-	proc, _ := os.FindProcess(pid)
-	if err := proc.Signal(syscall.SIGUSR1); err != nil {
-		log.Println(Type(c), Name(c), "refresh failed")
-
-	}
-	return
-}
 
 // locate a process by compoent type and name
 //
