--- cmd/geneosctl/geneosctl_windows.go
@@ -1,12 +1,0 @@
-//go:build windows
-package main
-
-var itrsHome string = "C:\\Program Files\\NetprobeNT"
-
-func refresh(ct ComponentType, pid int, name string) {
-	// do nothing
-}
-
-func findProc(c Component, name string) (int, error) {
-	return
-}
